refactor(reports): extract pure evaluateTierBenchmark into shared module
